fix(DatasetMetadataTip): rename to DatasetMetadataFieldTip
